ai_summarizer/requirements.txt
google-generativeai==0.8.5
pypdf==5.4.0
python-docx==1.1.2
python-dotenv==1.0.1
certifi==2025.1.31
